assignment07c/assignment07c/assignment07c/urls.py
from django.conf.urls import url
from django.contrib import admin
from django.urls import path, include
from rest_framework import routers
from rest_framework_swagger.views import get_swagger_view

from delivery import views

router = routers.DefaultRouter()
router.register(r'restaurants', views.RestaurantViewSet)
router.register(r'items', views.ItemViewSet)

schema_view = get_swagger_view(title='Delivery Site API')

urlpatterns = [
    path('admin/', admin.site.urls),
    url(r'api/', include(router.urls)),
    url(r'^$', schema_view)
]
